usbprod-common: Allow for USB partition resizing
Also moves LED blinking to blast_funcs.sh, and adds different script
hooks for blast.sh from S99blast to allow for this shift.

Partition resizing allows this process to output a very small .dd
file that can be written to USB drive from nearly any OS. It will
expand to the full disk length on the first run. And, since there
are no disk images present on the USB drive, will start the capture
process.

The goal of this is to greatly simplify the customer process to
capture and replicate images.